refactor(events): #1 イベント一覧カードのUIを改善しコンポーネントを分離
- CardHeader を廃止し CardContent のみのコンパクトなレイアウトに変更
- カード全体をリンクにし、右端に ChevronRight を配置（積立カードと統一）
- EventCard・EventCreateDialog を -components/ に切り出しページを薄く整理
